feat: support from iOS 10 and above
According to https://github.com/omise/omise-ios/pull/122
It said Omise iOS SDK is starting supporting iOS 10 and above.

And to follow to the Bank of Thailand (BOT). We should not allow the
user to use an obsolete version of operating systems (OS).

BTW. We must consider 12 and above now since Apple emphasize developer
to just support only latest version - 2.

And it challenging when we need to support more versions. 🤣