chore: setup pre-commit
by just use what I think we should use from the pre-commit-hooks that
the pre-commit project has proved.

And use some custom configs for SwiftLint and SwiftFormat.

To run SwiftLint with `--fix` command first is to use SwiftLint to fix
the problem. And use SwiftFormat to fix syntax more. Then check again
with SwiftLint and